fix: MEDIUM priority compliance — formatDate consolidation, comments, docs, hygiene
- formatDate: consolidated 3 defs to canonical src/lib/formatters.ts
  - test-tag-reminders.ts: import from @/lib/formatters
  - crew-emails.ts: import from @/lib/formatters
  - pdfme/helpers.ts: re-export from @/lib/formatters (PDF/UI alignment)
- Obvious WHAT comments: removed from kits.ts, stocktake.ts, categories.ts
- WHY comments: added to line-items.ts, warehouse.ts complex branches
- TODO hygiene: removed stale TODOs from damage/page.tsx, project-costs-panel.tsx
- console.log: env-guarded in email.ts, structured devLog in bot-process.ts
- line-item-fulfillment: moved from src/server/ to src/lib/ (library, not server action)
  - Updated 6 imports across warehouse.ts, bulk-checkin.ts, check-records.ts, etc.
- 02-project-structure.md: updated to reflect all ~70 server files
